backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.22
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3
pydantic==2.7.4
pydantic-settings==2.3.4
redis[hiredis]==5.0.7
minio==7.2.7
qdrant-client==1.9.2
PyMuPDF==1.24.5
Pillow==12.1.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
